requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.6
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
protobuf==4.21.9
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
